Refactor ESN training to a single file
There was a lot of repetition in my previous approach (a separate ESN
file for each solver method) when the only thing changing was the
training function.

This commit puts all training methods in one file and adds a selection
method to the BaseESN module.

The selection method takes a string for BenchmarkTools compatibility.